fix docker-compose examples UID/GID settings
when you change the PUID/PGID in the `.env` file, only the librenms
containers UID/GID are changes which causes permission issues with
rrdcached. rrdcached will still be running with UID=1000 and GID=1000.
As a result of that, no graphs will be rendered.

By adding:
```
- "PUID=${PUID}"
- "PGID=${PGID}"
```

to the rrdcache service the graphs will rendered properly because
rrdcached will also run with the same UID/GID as librenms

Signed-off-by: BlackEagle <ike.devolder@gmail.com>